fix
git-svn-id: svn://172.16.27.209/gsgh/gsgh/lghjft-backend@595 63aaa323-bd20-4782-bdcf-f624a29d9816

+++ yudao-module-lghjft/src/main/java/cn/iocoder/yudao/module/lghjft/controller/admin/xxzx/xxtx/vo/XxtxMessagePageReqVO.java
@@ -34,6 +34,6 @@ public class XxtxMessagePageReqVO extends PageParam {
     private Integer readStatus;
 
     @Schema(description = "接收着ID", example = "1")
-    private Integer receiverId;
+    private Long receiverId;
 
 }
